Remove internal use of sf::ConvexShape. * Replace with class Shape ** a simple wrapper around    an array of points

+++ GravityCommon/CommandQueue.cpp
@@ -18,7 +18,7 @@ bool CommandQueue::nextCommand(RigidBody & rb, Command & c)
 
 
 
-double CommandQueue::getTimestamp()
+uint64_t CommandQueue::getTimestamp()
 {
     FILETIME t;
     GetSystemTimePreciseAsFileTime(&t);
